OPEN - task 9: JavaEE7 CH-8 Object-Relational Mapping https://github.com/nilcy/zeroth/issues/issues/issue/9

+++ zeroth-actor/zeroth-actor-screen/src/main/java/zeroth/actor/screen/iface/jsf/conversation/home/NewsAction.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.screen.iface.jsf.conversation.home;
+import java.math.BigDecimal;
 import zeroth.actor.service.domain.misc.News;
 import zeroth.framework.screen.iface.jsf.Action;
 import zeroth.framework.standard.shared.SimpleFilter;
@@ -11,5 +12,5 @@ import zeroth.framework.standard.shared.SimpleFilter;
  * お知らせアクションI/F
  * @author nilcy
  */
-public interface NewsAction extends Action<News, Long, SimpleFilter> {
+public interface NewsAction extends Action<News, BigDecimal, SimpleFilter> {
 }
