+++ zeroth-actor/zeroth-actor-screen/src/main/java/zeroth/actor/screen/iface/jsf/conversation/master/IndustryClassAction.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.screen.iface.jsf.conversation.master;
+import java.math.BigDecimal;
 import javax.enterprise.context.ConversationScoped;
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -20,7 +21,8 @@ import zeroth.framework.standard.shared.SimpleFilter;
  */
 @Named(value = "industryClassAction")
 @ConversationScoped
-public class IndustryClassAction extends AbstractActionImpl<IndustryClass, Long, SimpleFilter> {
+public class IndustryClassAction extends
+    AbstractActionImpl<IndustryClass, BigDecimal, SimpleFilter> {
     /** 製品番号 */
     private static final long serialVersionUID = 7243039551620767571L;
     /** 業種アプリケーションI/F */
@@ -32,7 +34,7 @@ public class IndustryClassAction extends AbstractActionImpl<IndustryClass, Long,
     }
     /** {@inheritDoc} */
     @Override
-    public SimpleRepositoryApplication<IndustryClass, Long, SimpleFilter> getApplication() {
+    public SimpleRepositoryApplication<IndustryClass, BigDecimal, SimpleFilter> getApplication() {
         return industryClassApplication;
     }
     /** {@inheritDoc} */
