+++ zeroth-actor/zeroth-actor-screen/src/main/java/zeroth/actor/screen/iface/jsf/conversation/master/ListedSectionAction.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.screen.iface.jsf.conversation.master;
+import java.math.BigDecimal;
 import javax.enterprise.context.ConversationScoped;
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -20,7 +21,8 @@ import zeroth.framework.standard.shared.SimpleFilter;
  */
 @Named(value = "listedSectionAction")
 @ConversationScoped
-public class ListedSectionAction extends AbstractActionImpl<ListedSection, Long, SimpleFilter> {
+public class ListedSectionAction extends
+    AbstractActionImpl<ListedSection, BigDecimal, SimpleFilter> {
     /** 製品番号 */
     private static final long serialVersionUID = 7243039551620767571L;
     /** 上場先アプリケーションI/F */
@@ -32,7 +34,7 @@ public class ListedSectionAction extends AbstractActionImpl<ListedSection, Long,
     }
     /** {@inheritDoc} */
     @Override
-    public SimpleRepositoryApplication<ListedSection, Long, SimpleFilter> getApplication() {
+    public SimpleRepositoryApplication<ListedSection, BigDecimal, SimpleFilter> getApplication() {
         return listedSectionApplication;
     }
     /** {@inheritDoc} */
