+++ zeroth-actor/zeroth-actor-screen/src/main/java/zeroth/actor/screen/iface/jsf/conversation/master/SupplierAction.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.screen.iface.jsf.conversation.master;
+import java.math.BigDecimal;
 import javax.enterprise.context.ConversationScoped;
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -22,7 +23,7 @@ import zeroth.framework.standard.shared.SimpleFilter;
  */
 @Named(value = "supplierAction")
 @ConversationScoped
-public class SupplierAction extends AbstractActionImpl<Supplier, Long, SimpleFilter> {
+public class SupplierAction extends AbstractActionImpl<Supplier, BigDecimal, SimpleFilter> {
     /** 製品番号 */
     private static final long serialVersionUID = 873776474936603723L;
     /** 調達先アプリケーションI/F */
@@ -34,7 +35,7 @@ public class SupplierAction extends AbstractActionImpl<Supplier, Long, SimpleFil
     }
     /** {@inheritDoc} */
     @Override
-    public SimpleRepositoryApplication<Supplier, Long, SimpleFilter> getApplication() {
+    public SimpleRepositoryApplication<Supplier, BigDecimal, SimpleFilter> getApplication() {
         return supplierApplication;
     }
     /** {@inheritDoc} */
