+++ zeroth-actor/zeroth-actor-service-api/src/main/java/zeroth/actor/service/app/actor/MemberApplication.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.service.app.actor;
+import java.math.BigDecimal;
 import javax.ejb.Local;
 import zeroth.actor.service.domain.Member;
 import zeroth.actor.service.domain.MemberFilter;
@@ -13,5 +14,6 @@ import zeroth.framework.enterprise.app.SimpleRepositoryApplication;
  * @author nilcy
  */
 @Local
-public interface MemberApplication extends SimpleRepositoryApplication<Member, Long, MemberFilter> {
+public interface MemberApplication extends
+    SimpleRepositoryApplication<Member, BigDecimal, MemberFilter> {
 }
