+++ zeroth-actor/zeroth-actor-service-api/src/main/java/zeroth/actor/service/app/misc/IndustryClassApplication.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.service.app.misc;
+import java.math.BigDecimal;
 import javax.ejb.Local;
 import zeroth.actor.service.domain.misc.IndustryClass;
 import zeroth.framework.enterprise.app.SimpleRepositoryApplication;
@@ -14,5 +15,5 @@ import zeroth.framework.standard.shared.SimpleFilter;
  */
 @Local
 public interface IndustryClassApplication extends
-    SimpleRepositoryApplication<IndustryClass, Long, SimpleFilter> {
+    SimpleRepositoryApplication<IndustryClass, BigDecimal, SimpleFilter> {
 }
