+++ zeroth-actor/zeroth-actor-service/src/main/java/zeroth/actor/service/app/actor/MemberApplicationImpl.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.service.app.actor;
+import java.math.BigDecimal;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import zeroth.actor.service.domain.Member;
@@ -20,7 +21,8 @@ import zeroth.framework.enterprise.shared.Tracer;
 // @Default
 @Tracer
 public class MemberApplicationImpl extends
-    AbstractSimpleRepositoryApplication<Member, Long, MemberFilter> implements MemberApplication {
+    AbstractSimpleRepositoryApplication<Member, BigDecimal, MemberFilter> implements
+    MemberApplication {
     /** 製品番号 */
     private static final long serialVersionUID = 2537651945740718957L;
     /** 社員リポジトリ */
@@ -31,7 +33,7 @@ public class MemberApplicationImpl extends
     }
     /** {@inheritDoc} */
     @Override
-    protected SimpleRepository<Member, Long, MemberFilter> getRepository() {
+    protected SimpleRepository<Member, BigDecimal, MemberFilter> getRepository() {
         return repository;
     }
 }
