+++ zeroth-actor/zeroth-actor-service/src/main/java/zeroth/actor/service/app/actor/OrganApplicationImpl.java
@@ -4,6 +4,7 @@
 // http://www.gnu.org/licenses/agpl-3.0.txt
 // ========================================================================
 package zeroth.actor.service.app.actor;
+import java.math.BigDecimal;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import zeroth.actor.service.domain.Organ;
@@ -20,7 +21,8 @@ import zeroth.framework.standard.shared.SimpleFilter;
 // @Default
 @Tracer
 public class OrganApplicationImpl extends
-    AbstractSimpleRepositoryApplication<Organ, Long, SimpleFilter> implements OrganApplication {
+    AbstractSimpleRepositoryApplication<Organ, BigDecimal, SimpleFilter> implements
+    OrganApplication {
     /** 製品番号 */
     private static final long serialVersionUID = 2537651945740718957L;
     /** 組織リポジトリ */
@@ -31,7 +33,7 @@ public class OrganApplicationImpl extends
     }
     /** {@inheritDoc} */
     @Override
-    protected SimpleRepository<Organ, Long, SimpleFilter> getRepository() {
+    protected SimpleRepository<Organ, BigDecimal, SimpleFilter> getRepository() {
         return repository;
     }
 }
